Split DaemonRpc up into two parts
